Minor test cleanups (create some new fixtures)
